backend/requirements.txt
asgiref==3.7.2
djoser==2.1.0
Django==3.2
django-filter==22.1
djangorestframework==3.14.0
drf-extra-fields==3.4.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
Pillow==10.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
